WIP: conslidate node reachability check in one pkg
EgressIP and EgressSVC both will be the consumer of this
new package
