+++ go-controller/pkg/ovn/default_network_controller.go
@@ -194,9 +194,7 @@ func newDefaultNetworkControllerCommon(cnci *CommonNetworkControllerInfo,
 			allocator:                         allocator{&sync.Mutex{}, make(map[string]*egressNode)},
 			nbClient:                          cnci.nbClient,
 			watchFactory:                      cnci.watchFactory,
-			egressIPTotalTimeout:              config.OVNKubernetesFeature.EgressIPReachabiltyTotalTimeout,
 			reachabilityCheckInterval:         egressIPReachabilityCheckInterval,
-			egressIPNodeHealthCheckPort:       config.OVNKubernetesFeature.EgressIPNodeHealthCheckPort,
 		},
 		loadbalancerClusterCache: make(map[kapi.Protocol]string),
 		loadBalancerGroupUUID:    "",
