+++ go-controller/pkg/ovn/egressip.go
@@ -1,18 +1,15 @@
 package ovn
 
 import (
-	"context"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"net"
-	"os"
 	"reflect"
 	"sort"
 	"strings"
 	"sync"
-	"syscall"
 	"time"
 
 	ocpcloudnetworkapi "github.com/openshift/api/cloudnetwork/v1"
@@ -24,6 +21,7 @@ import (
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdbops"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/metrics"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nbdb"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nodereachability"
 	addressset "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/address_set"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/healthcheck"
 
@@ -40,12 +38,6 @@ import (
 	"k8s.io/klog/v2"
 	utilnet "k8s.io/utils/net"
 )
-
-type egressIPDialer interface {
-	dial(ip net.IP, timeout time.Duration) bool
-}
-
-var dialer egressIPDialer = &egressIPDial{}
 
 type healthcheckClientAllocator interface {
 	allocate(nodeName string) healthcheck.EgressIPHealthClient
@@ -1319,7 +1311,7 @@ func (oc *DefaultNetworkController) isEgressNodeReachable(egressNode *kapi.Node)
 	oc.eIPC.allocator.Lock()
 	defer oc.eIPC.allocator.Unlock()
 	if eNode, exists := oc.eIPC.allocator.cache[egressNode.Name]; exists {
-		return eNode.isReachable || oc.isReachable(eNode.name, eNode.mgmtIPs, eNode.healthClient)
+		return eNode.isReachable || nodereachability.IsReachable(eNode.name, eNode.mgmtIPs, eNode.healthClient)
 	}
 	return false
 }
@@ -2073,7 +2065,7 @@ func (oc *DefaultNetworkController) initClusterEgressPolicies(nodes []interface{
 		}
 	}
 
-	go oc.checkEgressNodesReachability()
+	go nodereachability.CheckNodesReachability(oc.stopChan, oc.checkEgressNodesReachabilityIterate)
 	return nil
 }
 
@@ -2196,12 +2188,8 @@ type egressIPController struct {
 	nbClient libovsdbclient.Client
 	// watchFactory watching k8s objects
 	watchFactory *factory.WatchFactory
-	// EgressIP Node reachability total timeout configuration
-	egressIPTotalTimeout int
 	// reachability check interval
 	reachabilityCheckInterval time.Duration
-	// EgressIP Node reachability gRPC port (0 means it should use dial instead)
-	egressIPNodeHealthCheckPort int
 }
 
 // addStandByEgressIPAssignment does the same setup that is done by addPodEgressIPAssignments but for
@@ -2566,31 +2554,13 @@ func (oc *DefaultNetworkController) deletePodIPsFromAddressSet(addrSetIPs []net.
 	return nil
 }
 
-// checkEgressNodesReachability continuously checks if all nodes used for egress
-// IP assignment are reachable, and updates the nodes following the result. This
-// is important because egress IP is based upon routing traffic to these nodes,
-// and if they aren't reachable we shouldn't be using them for egress IP.
-func (oc *DefaultNetworkController) checkEgressNodesReachability() {
-	timer := time.NewTicker(oc.eIPC.reachabilityCheckInterval)
-	defer timer.Stop()
-	for {
-		select {
-		case <-timer.C:
-			checkEgressNodesReachabilityIterate(oc)
-		case <-oc.stopChan:
-			klog.V(5).Infof("Stop channel got triggered: will stop checkEgressNodesReachability")
-			return
-		}
-	}
-}
-
-func checkEgressNodesReachabilityIterate(oc *DefaultNetworkController) {
+func (oc *DefaultNetworkController) checkEgressNodesReachabilityIterate() {
 	reAddOrDelete := map[string]bool{}
 	oc.eIPC.allocator.Lock()
 	for _, eNode := range oc.eIPC.allocator.cache {
 		if eNode.isEgressAssignable && eNode.isReady {
 			wasReachable := eNode.isReachable
-			isReachable := oc.isReachable(eNode.name, eNode.mgmtIPs, eNode.healthClient)
+			isReachable := nodereachability.IsReachable(eNode.name, eNode.mgmtIPs, eNode.healthClient)
 			if wasReachable && !isReachable {
 				reAddOrDelete[eNode.name] = true
 			} else if !wasReachable && isReachable {
@@ -2622,99 +2592,10 @@ func checkEgressNodesReachabilityIterate(oc *DefaultNetworkController) {
 	}
 }
 
-func (oc *DefaultNetworkController) isReachable(nodeName string, mgmtIPs []net.IP, healthClient healthcheck.EgressIPHealthClient) bool {
-	// Check if we need to do node reachability check
-	if oc.eIPC.egressIPTotalTimeout == 0 {
-		return true
-	}
-
-	if oc.eIPC.egressIPNodeHealthCheckPort == 0 {
-		return isReachableLegacy(nodeName, mgmtIPs, oc.eIPC.egressIPTotalTimeout)
-	}
-	return isReachableViaGRPC(mgmtIPs, healthClient, oc.eIPC.egressIPNodeHealthCheckPort, oc.eIPC.egressIPTotalTimeout)
-}
-
-func isReachableLegacy(node string, mgmtIPs []net.IP, totalTimeout int) bool {
-	var retryTimeOut, initialRetryTimeOut time.Duration
-
-	numMgmtIPs := len(mgmtIPs)
-	if numMgmtIPs == 0 {
-		return false
-	}
-
-	switch totalTimeout {
-	// Check if we need to do node reachability check
-	case 0:
-		return true
-	case 1:
-		// Using time duration for initial retry with 700/numIPs msec and retry of 100/numIPs msec
-		// to ensure total wait time will be in range with the configured value including a sleep of 100msec between attempts.
-		initialRetryTimeOut = time.Duration(700/numMgmtIPs) * time.Millisecond
-		retryTimeOut = time.Duration(100/numMgmtIPs) * time.Millisecond
-	default:
-		// Using time duration for initial retry with 900/numIPs msec
-		// to ensure total wait time will be in range with the configured value including a sleep of 100msec between attempts.
-		initialRetryTimeOut = time.Duration(900/numMgmtIPs) * time.Millisecond
-		retryTimeOut = initialRetryTimeOut
-	}
-
-	timeout := initialRetryTimeOut
-	endTime := time.Now().Add(time.Second * time.Duration(totalTimeout))
-	for time.Now().Before(endTime) {
-		for _, ip := range mgmtIPs {
-			if dialer.dial(ip, timeout) {
-				return true
-			}
-		}
-		time.Sleep(100 * time.Millisecond)
-		timeout = retryTimeOut
-	}
-	klog.Errorf("Failed reachability check for %s", node)
-	return false
-}
-
-type egressIPDial struct{}
-
-// Blantant copy from: https://github.com/openshift/sdn/blob/master/pkg/network/common/egressip.go#L499-L505
-// Ping a node and return whether or not we think it is online. We do this by trying to
-// open a TCP connection to the "discard" service (port 9); if the node is offline, the
-// attempt will either time out with no response, or else return "no route to host" (and
-// we will return false). If the node is online then we presumably will get a "connection
-// refused" error; but the code below assumes that anything other than timeout or "no
-// route" indicates that the node is online.
-func (e *egressIPDial) dial(ip net.IP, timeout time.Duration) bool {
-	conn, err := net.DialTimeout("tcp", net.JoinHostPort(ip.String(), "9"), timeout)
-	if conn != nil {
-		conn.Close()
-	}
-	if opErr, ok := err.(*net.OpError); ok {
-		if opErr.Timeout() {
-			return false
-		}
-		if sysErr, ok := opErr.Err.(*os.SyscallError); ok && sysErr.Err == syscall.EHOSTUNREACH {
-			return false
-		}
-	}
-	return true
-}
-
 type egressIPHealthcheckClientAllocator struct{}
 
 func (hccAlloc *egressIPHealthcheckClientAllocator) allocate(nodeName string) healthcheck.EgressIPHealthClient {
 	return healthcheck.NewEgressIPHealthClient(nodeName)
-}
-
-func isReachableViaGRPC(mgmtIPs []net.IP, healthClient healthcheck.EgressIPHealthClient, healthCheckPort, totalTimeout int) bool {
-	dialCtx, dialCancel := context.WithTimeout(context.Background(), time.Duration(totalTimeout)*time.Second)
-	defer dialCancel()
-
-	if !healthClient.IsConnected() {
-		// gRPC session is not up. Attempt to connect and if that suceeds, we will declare node as reacheable.
-		return healthClient.Connect(dialCtx, mgmtIPs, healthCheckPort)
-	}
-
-	// gRPC session is already established. Send a probe, which will succeed, or close the session.
-	return healthClient.Probe(dialCtx)
 }
 
 func getClusterSubnets() ([]*net.IPNet, []*net.IPNet) {
