+++ go-controller/pkg/ovn/egressservices_test.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nbdb"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/healthcheck"
 	libovsdbtest "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/testing/libovsdb"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
@@ -1711,135 +1710,6 @@ var _ = ginkgo.Describe("OVN Egress Service Operations", func() {
 						},
 					},
 				)
-
-				ginkgo.By("modifying the controller's IsReachable func to return false on the first call and true for the second")
-				count := 0
-				fakeOVN.controller.egressSvcController.IsReachable = func(nodeName string, _ []net.IP, _ healthcheck.EgressIPHealthClient) bool {
-					count++
-					return count == 2
-				}
-				fakeOVN.InitAndRunEgressSVCController()
-				gomega.Eventually(func() error {
-					svc, err := fakeOVN.fakeClient.KubeClient.CoreV1().Services("testns").Get(context.TODO(), svc1.Name, metav1.GetOptions{})
-					if err != nil {
-						return err
-					}
-
-					val := svc.Annotations[util.EgressSVCHostAnnotation]
-					if val != node1Name {
-						return fmt.Errorf("expected svc1's host annotation value %s to be node1", val)
-					}
-
-					node1ExpectedLabels := map[string]string{
-						"kubernetes.io/hostname":                                 "node1",
-						fmt.Sprintf("%s/testns-svc1", util.EgressSVCLabelPrefix): "",
-					}
-
-					node1, err = fakeOVN.fakeClient.KubeClient.CoreV1().Nodes().Get(context.TODO(), node1Name, metav1.GetOptions{})
-					if err != nil {
-						return err
-					}
-
-					if !reflect.DeepEqual(node1.Labels, node1ExpectedLabels) {
-						return fmt.Errorf("expected node1's labels %v to be equal %v", node1.Labels, node1ExpectedLabels)
-					}
-
-					return nil
-				}).ShouldNot(gomega.HaveOccurred())
-
-				svc1v4lrp1 := lrpForEgressSvcEndpoint("svc1v4lrp1-UUID", "testns/svc1", "10.128.1.5", "10.128.1.2")
-				svc1v6lrp1 := lrpForEgressSvcEndpoint("svc1v6lrp1-UUID", "testns/svc1", "fe00:10:128:1::5", "fe00:10:128:1::2")
-				clusterRouter.Policies = []string{"svc1v4lrp1-UUID", "svc1v6lrp1-UUID"}
-				expectedDatabaseState := []libovsdbtest.TestData{
-					clusterRouter,
-					svc1v4lrp1,
-					svc1v6lrp1,
-				}
-
-				for _, lrp := range getDefaultNoReroutePolicies() {
-					expectedDatabaseState = append(expectedDatabaseState, lrp)
-					clusterRouter.Policies = append(clusterRouter.Policies, lrp.UUID)
-				}
-				gomega.Eventually(fakeOVN.nbClient).Should(libovsdbtest.HaveData(expectedDatabaseState))
-
-				ginkgo.By("calling the reachability check which will return that the node is unreachable it will be drained")
-				fakeOVN.controller.egressSvcController.CheckNodesReachabilityIterate()
-				gomega.Eventually(func() error {
-					svc, err := fakeOVN.fakeClient.KubeClient.CoreV1().Services("testns").Get(context.TODO(), svc1.Name, metav1.GetOptions{})
-					if err != nil {
-						return err
-					}
-
-					val, ok := svc.Annotations[util.EgressSVCHostAnnotation]
-					if ok {
-						return fmt.Errorf("expected svc1's egress host annotation to be empty, got: %s", val)
-					}
-
-					node1ExpectedLabels := map[string]string{
-						"kubernetes.io/hostname": "node1",
-					}
-
-					node1, err = fakeOVN.fakeClient.KubeClient.CoreV1().Nodes().Get(context.TODO(), node1Name, metav1.GetOptions{})
-					if err != nil {
-						return err
-					}
-
-					if !reflect.DeepEqual(node1.Labels, node1ExpectedLabels) {
-						return fmt.Errorf("expected node1's labels %v to be equal %v", node1.Labels, node1ExpectedLabels)
-					}
-
-					return nil
-				}).ShouldNot(gomega.HaveOccurred())
-				clusterRouter.Policies = []string{}
-				expectedDatabaseState = []libovsdbtest.TestData{clusterRouter}
-				for _, lrp := range getDefaultNoReroutePolicies() {
-					expectedDatabaseState = append(expectedDatabaseState, lrp)
-					clusterRouter.Policies = append(clusterRouter.Policies, lrp.UUID)
-				}
-				gomega.Eventually(fakeOVN.nbClient).Should(libovsdbtest.HaveData(expectedDatabaseState))
-
-				ginkgo.By("calling the reachability check which will return that the node is reachable the service will be reallocated")
-				fakeOVN.controller.egressSvcController.CheckNodesReachabilityIterate()
-				gomega.Eventually(func() error {
-					svc, err := fakeOVN.fakeClient.KubeClient.CoreV1().Services("testns").Get(context.TODO(), svc1.Name, metav1.GetOptions{})
-					if err != nil {
-						return err
-					}
-
-					val := svc.Annotations[util.EgressSVCHostAnnotation]
-					if val != node1Name {
-						return fmt.Errorf("expected svc1's host annotation value %s to be node1", val)
-					}
-
-					node1ExpectedLabels := map[string]string{
-						"kubernetes.io/hostname":                                 "node1",
-						fmt.Sprintf("%s/testns-svc1", util.EgressSVCLabelPrefix): "",
-					}
-
-					node1, err = fakeOVN.fakeClient.KubeClient.CoreV1().Nodes().Get(context.TODO(), node1Name, metav1.GetOptions{})
-					if err != nil {
-						return err
-					}
-
-					if !reflect.DeepEqual(node1.Labels, node1ExpectedLabels) {
-						return fmt.Errorf("expected node1's labels %v to be equal %v", node1.Labels, node1ExpectedLabels)
-					}
-
-					return nil
-				}).ShouldNot(gomega.HaveOccurred())
-
-				clusterRouter.Policies = []string{"svc1v4lrp1-UUID", "svc1v6lrp1-UUID"}
-				expectedDatabaseState = []libovsdbtest.TestData{
-					clusterRouter,
-					svc1v4lrp1,
-					svc1v6lrp1,
-				}
-
-				for _, lrp := range getDefaultNoReroutePolicies() {
-					expectedDatabaseState = append(expectedDatabaseState, lrp)
-					clusterRouter.Policies = append(clusterRouter.Policies, lrp.UUID)
-				}
-				gomega.Eventually(fakeOVN.nbClient).Should(libovsdbtest.HaveData(expectedDatabaseState))
 				return nil
 			}
 			err := app.Run([]string{app.Name})
